pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "assistant-core"
version = "0.1.0"
description = "Modular personal AI assistant with Telegram, skills, and local/cloud models"
readme = "assistant/README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "openai>=1.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]
dashboard = [
    "flask>=3.0.0",
]

[project.scripts]
assistant-core = "assistant.main:main"
telegram-adapter = "assistant.channels.telegram:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["assistant*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["assistant/tests"]
pythonpath = ["."]
